+++ udata/commands/fixtures.py
@@ -3,7 +3,9 @@ from __future__ import unicode_literals
 
 import logging
 
-from udata.commands import manager
+import click
+
+from udata.commands import cli
 from udata.core.dataset.factories import VisibleDatasetFactory
 from udata.core.discussions.factories import DiscussionFactory
 from udata.core.organization.factories import OrganizationFactory
@@ -14,8 +16,12 @@ from udata.core.user.factories import UserFactory
 log = logging.getLogger(__name__)
 
 
-@manager.command
-def generate_fixtures(datasets=5, reuses=1):
+@cli.command()
+@click.option('-d', '--datasets', type=int, default=5,
+              help='Number of datasets to generate')
+@click.option('-r', '--reuses', type=int, default=1,
+              help='Number of reuses by dataset')
+def generate_fixtures(datasets, reuses):
     '''Build sample fixture data (users, datasets and reuses).'''
     user = UserFactory()
     log.info('Generated user "{user.email}".'.format(user=user))