+++ udata/commands/init.py
@@ -3,12 +3,12 @@ from __future__ import unicode_literals
 
 import logging
 
-from flask_script import prompt_bool
+import click
 
-from udata.commands import manager, IS_INTERACTIVE, green
+from udata.commands import cli, success, IS_TTY
 from udata.core.dataset.commands import licenses
 from udata.core.user import commands as user_commands
-from udata.i18n import lazy_gettext as _
+from udata.i18n import gettext as _
 from udata.search.commands import index
 
 from .db import migrate
@@ -17,27 +17,28 @@ from .fixtures import generate_fixtures
 log = logging.getLogger(__name__)
 
 
-@manager.command
-def init():
+@cli.command()
+@click.pass_context
+def init(ctx):
     '''Initialize your udata instance (search index, user, sample data...)'''
 
     log.info('Apply DB migrations if needed')
-    migrate(record=True)
+    ctx.invoke(migrate, record=True)
 
-    index()
+    ctx.invoke(index)
 
-    if IS_INTERACTIVE:
+    if IS_TTY:
         text = _('Do you want to create a superadmin user?')
-        if prompt_bool(text, default=True):
-            user = user_commands.create()
-            user_commands.set_admin(user.email)
+        if click.confirm(text, default=True):
+            user = ctx.invoke(user_commands.create)
+            ctx.invoke(user_commands.set_admin, email=user.email)
 
         text = _('Do you want to import some data-related license?')
-        if prompt_bool(text, default=True):
-            licenses()
+        if click.confirm(text, default=True):
+            ctx.invoke(licenses)
 
         text = _('Do you want to create some sample data?')
-        if prompt_bool(text, default=True):
-            generate_fixtures()
+        if click.confirm(text, default=True):
+            ctx.invoke(generate_fixtures)
 
-    log.info(green(_('Your udata instance is ready!')))
+    success(_('Your udata instance is ready!'))