+++ udata/commands/static.py
@@ -5,29 +5,29 @@ import logging
 import os
 import shutil
 
+import click
+
 from glob import iglob
 from os import makedirs
 from os.path import exists, join, isdir
-from sys import exit
 
 from flask import current_app
-from flask_script import prompt_bool
 
-from udata.commands import manager
+from udata.commands import cli
 
 log = logging.getLogger(__name__)
 
 
-@manager.option('path', nargs='?', default='static', help='target path')
-@manager.option('-ni', '--no-input', dest="input",
-                action='store_false', help="Disable input prompts")
-def collect(path, input):
+@cli.command()
+@click.argument('path', default='public')
+@click.option('-ni', '--no-input', is_flag=True, help="Disable input prompts")
+def collect(path, no_input):
     '''Collect static files'''
     if exists(path):
         msg = '"%s" directory already exists and will be erased'
         log.warning(msg, path)
-        if input and not prompt_bool('Are you sure'):
-            exit(-1)
+        if not no_input:
+            click.confirm('Are you sure?', abort=True)
 
         log.info('Deleting static directory "%s"', path)
         shutil.rmtree(path)
@@ -51,7 +51,7 @@ def collect(path, input):
             destination = join(path, prefix)
             copy_recursive(blueprint.static_folder, destination)
 
-    for prefix, source in manager.app.config['STATIC_DIRS']:
+    for prefix, source in current_app.config['STATIC_DIRS']:
         log.info('Copying %s to %s', source, prefix)
         destination = join(path, prefix)
         copy_recursive(source, destination)