+++ udata/commands/worker.py
@@ -3,36 +3,39 @@ from __future__ import unicode_literals
 
 import json
 import logging
-import sys
 from collections import Counter
 from urlparse import urlparse
 
+import click
 import redis
 
 from celery.task.control import inspect
 from flask import current_app
 
 from udata.app import cache
-from udata.commands import submanager, red
+from udata.commands import cli, exit_with_error
 from udata.tasks import celery
 
 log = logging.getLogger(__name__)
 
-m = submanager(
-    'worker',
-    help='Worker related operations',
-    description='Handle all worker related operations and maintenance'
-)
+
+@cli.group('worker')
+def grp():
+    '''Worker related operations'''
+    pass
+
 
 TASKS_LIST_CACHE_KEY = 'worker-status-tasks'
 # we're using an aggressive cache in order not to hit Celery every 5 min
-TASKS_LIST_CACHE_DURATION = 60*60*24  # in seconds
+TASKS_LIST_CACHE_DURATION = 60 * 60 * 24  # in seconds
 
 
-@m.command
+@grp.command()
 def start():
-    """Start a worker"""
-    celery.start()
+    '''Start a worker'''
+    worker = celery.Worker()
+    worker.start()
+    return worker.exitcode
 
 
 def status_print_task(count, biggest_task_name, munin=False):
@@ -44,8 +47,7 @@ def status_print_task(count, biggest_task_name, munin=False):
 
 def status_print_config(queue):
     if not queue:
-        print(red('--munin-config called without a --queue parameter'))
-        sys.exit(-1)
+        exit_with_error('--munin-config called without a --queue parameter')
     tasks = cache.get(TASKS_LIST_CACHE_KEY) or []
     if not tasks:
         registered = inspect().registered_tasks()
@@ -89,8 +91,7 @@ def get_queues(queue):
     if queue:
         queues = [q for q in queues if q == queue]
     if not len(queues):
-        print(red('Error: no queue found'))
-        sys.exit(-1)
+        exit_with_error('Error: no queue found')
     return queues
 
 
@@ -101,11 +102,12 @@ def get_redis_connection():
                              db=db)
 
 
-@m.option('-q', '--queue', help='Queue to be analyzed', default=None)
-@m.option('-m', '--munin', action='store_true', dest='munin',
-          help='Output in a munin plugin compatible format')
-@m.option('-c', '--munin-config', action='store_true', dest='munin_config',
-          help='Output in a munin plugin config compatible format')
+@grp.command()
+@click.option('-q', '--queue', help='Queue to be analyzed', default=None)
+@click.option('-m', '--munin', is_flag=True,
+              help='Output in a munin plugin compatible format')
+@click.option('-c', '--munin-config', is_flag=True,
+              help='Output in a munin plugin config compatible format')
 def status(queue, munin, munin_config):
     """List queued tasks aggregated by name"""
     if munin_config: