+++ udata/core/badges/commands.py
@@ -1,21 +1,20 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
+import click
 import logging
 
 from os.path import exists
 
-from udata.commands import submanager
+from udata.commands import cli
 from udata.models import Organization
 
 log = logging.getLogger(__name__)
 
 
-m = submanager(
-    'badges',
-    help='Badges related operations',
-    description='Handle all badges related operations and maintenance'
-)
+@cli.group('badges')
+def grp():
+    '''Badges related operations'''
 
 
 def toggle_badge(id_or_slug, badge_kind):
@@ -30,7 +29,9 @@ def toggle_badge(id_or_slug, badge_kind):
     organization.toggle_badge(badge_kind)
 
 
-@m.command
+@grp.command()
+@click.argument('path_or_id')
+@click.argument('badge_kind')
 def toggle(path_or_id, badge_kind):
     '''Toggle a `badge_kind` for a given `path_or_id`