+++ udata/core/organization/commands.py
@@ -3,21 +3,25 @@ from __future__ import unicode_literals
 
 import logging
 
-from udata.commands import submanager
+import click
+
+from udata.commands import cli, exit_with_error
 from udata.models import GeoZone, Organization
 
 log = logging.getLogger(__name__)
 
-m = submanager(
-    'organizations',
-    help='Organizations related operations',
-    description='Handle all organizations related operations and maintenance'
-)
+
+@cli.group('organizations')
+def grp():
+    '''Organizations related operations'''
+    pass
 
 
-@m.command
+@grp.command()
+@click.argument('geoid', metavar='<geoid>')
+@click.argument('organization_id_or_slug', metavar='<organization>')
 def attach_zone(geoid, organization_id_or_slug):
-    '''Attach a zone restricted to level for a given organization.'''
+    '''Attach a zone <geoid> restricted to level for a given <organization>.'''
     organization = Organization.objects.get_by_id_or_slug(
         organization_id_or_slug)
     if not organization:
@@ -32,13 +36,16 @@ def attach_zone(geoid, organization_id_or_slug):
     log.info('Done')
 
 
-@m.command
+@grp.command()
+@click.argument('organization_id_or_slug', metavar='<organization>')
 def detach_zone(organization_id_or_slug):
-    '''Detach the zone of a given organization.'''
+    '''Detach the zone of a given <organization>.'''
     organization = Organization.objects.get_by_id_or_slug(
         organization_id_or_slug)
     if not organization:
-        log.error('No organization found for %s', organization_id_or_slug)
+        exit_with_error(
+            'No organization found for {0}'.format(organization_id_or_slug)
+        )
     log.info('Detaching {organization} from {organization.zone}'.format(
              organization=organization))
     organization.zone = None